move func def into cpp; update output_mesh
